+++ src/main/java/duke/DukeApp.java
@@ -1,5 +1,4 @@
-import java.util.ArrayList;
-import java.util.Scanner;
+package duke;
 
 
 public class DukeApp {
