+++ src/main/java/duke/EmptyDescriptionException.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class EmptyDescriptionException extends Exception
 {
     public EmptyDescriptionException(String message)
