+++ src/main/java/duke/Event.java
@@ -1,34 +1,36 @@
+package duke;
+
 import java.io.BufferedReader;
 import java.io.FileWriter;
 import java.io.IOException;
 
-public class Deadline extends Task{
-    protected String by;
-    public Deadline (String taskName, boolean taskDone, String by)
+public class Event extends Task{
+        protected String at;
+        public Event (String taskName, boolean taskDone, String at)
         {
             super(taskName, taskDone); // calls the parent constructor
-            this.by = by;
+            this.at = at;
         }
 
-    public Deadline() {
+    public Event() {
         super();
     }
 
     public String toString()
-    {
-        return "[D]" + super.toString() + " (by: "+ by + ")";
-    }
+        {
+            return "[E]" + super.toString() + " (at: "+ at + ")";
+        }
+
     public void write(FileWriter storage) throws IOException {
-        storage.write("D\n"); //to represent as todo
+        storage.write("E\n"); //to represent as todo
         super.write(storage);
-        storage.write(by + "\n");
+        storage.write(at + "\n");
     }
 
     public void read(BufferedReader fileRead) throws IOException
     {
         super.read(fileRead);
-        by = fileRead.readLine();
+        at = fileRead.readLine();
     }
 
-    }
-
+}
