+++ src/main/java/duke/InvalidCommandException.java
@@ -1,2 +1,4 @@
+package duke;
+
 public class InvalidCommandException extends Exception {
 }
