+++ src/main/java/duke/Serializable.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.io.BufferedReader;
 import java.io.FileReader;
 import java.io.FileWriter;
