+++ src/main/java/duke/Task.java
@@ -1,5 +1,6 @@
+package duke;
+
 import java.io.BufferedReader;
-import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
