+++ src/main/java/duke/Todo.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.io.FileWriter;
 import java.io.IOException;
